HW9/pyproject.toml
[project]
name = "hw9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "onnx>=1.20.0",
    "pillow>=12.0.0",
]
